Controversially, Add SBCameraScroll support
Since I've completely revamped the snow system, SBCameraScroll kinda wont work well with this. I'd make a PR on SBCameraScroll's end but I'm not sure how I'd hook into regionkit. So I've went ahead and fixed it on our end. No extra library required, if the mod is installed then it just loads the assembly, grabs a few fields and methods, and uses them where appropriate. If this is not something regionkit should have, feel free to remove it.